pyproject.toml
[tool.poetry]
name = "todo-ai-chatbot"
version = "1.0.0"
description = "AI-powered todo chatbot using OpenAI Agents SDK and MCP"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
openai = "^1.3.7"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.3"
alembic = "^1.12.1"
mcp-sdk = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
flake8 = "^6.1.0"
black = "^23.10.1"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"